requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
torch
langchain
pdfplumber
python-docx
nltk
scikit-learn
numpy
pandas
uvicorn
fastapi
requests
tqdm
groq
python-dotenv
